+++ opentelemetry-api/tests/trace/test_tracer.py
@@ -30,7 +30,7 @@ class TestTracer(unittest.TestCase):
             self.assertIsInstance(span, trace.Span)
 
     def test_use_span(self):
-        span = trace.DefaultSpan(trace.INVALID_SPAN_CONTEXT)
+        span = trace.NonRecordingSpan(trace.INVALID_SPAN_CONTEXT)
         with self.tracer.use_span(span):
             pass